Move ipns function to ipfs
