+++ lib/modules/ens/embarkjs.js
@@ -247,7 +247,3 @@ __embarkENS.registerSubDomain = function (name, address, callback) {
 __embarkENS.isAvailable = function () {
   return Boolean(this.isAvailable);
 };
-
-__embarkENS.register = function () {
-  return new Error("Not available with ENS");
-};
