+++ lib/modules/ipfs/index.js
@@ -19,12 +19,8 @@ class IPFS {
     this.webServerConfig = embark.config.webServerConfig;
     this.blockchainConfig = embark.config.blockchainConfig;
 
-    if (this.isIpfsStorageEnabledInTheConfig() || this.isIpfsNameEnabledInTheConfig()) {
-      this.downloadIpfsApi();
-      this.addDefaultToEmbarkJS();
-    }
-
     if (this.isIpfsStorageEnabledInTheConfig()) {
+      this.downloadIpfsApi();
       this.setServiceCheck();
       this.addStorageProviderToEmbarkJS();
       this.addObjectToConsole();
@@ -37,11 +33,6 @@ class IPFS {
         self.logger.info("IPFS node not found, attempting to start own node");
         self.startProcess(() => {});
       });
-    }
-
-    if (this.isIpfsNameEnabledInTheConfig()) {
-      this.addNamesystemProviderToEmbarkJS();
-      this.setNamesystemProvider();
     }
   }
 
@@ -102,25 +93,10 @@ class IPFS {
     utils.getJson(url, cb);
   }
 
-  addDefaultToEmbarkJS() {
-    let code = "";
-    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs' , 'default.js')).toString();
-
-    this.embark.addCodeToEmbarkJS(code);
-  }
-
   addStorageProviderToEmbarkJS() {
     let code = "";
-    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs', 'storage.js')).toString();
+    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs.js')).toString();
     code += "\nEmbarkJS.Storage.registerProvider('ipfs', __embarkIPFS);";
-
-    this.embark.addCodeToEmbarkJS(code);
-  }
-
-  addNamesystemProviderToEmbarkJS() {
-    let code = "";
-    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs', 'name.js')).toString();
-    code += "\nEmbarkJS.Names.registerProvider('ipns', __embarkIPFS);";
 
     this.embark.addCodeToEmbarkJS(code);
   }
@@ -160,23 +136,6 @@ class IPFS {
     let {enabled, available_providers, dappConnection} = this.storageConfig;
     return enabled && (available_providers.indexOf('ipfs') > 0 || dappConnection.find(c => c.provider === 'ipfs'));
   }
-
-  isIpfsNameEnabledInTheConfig() {
-    let {enabled, available_providers} = this.namesystemConfig;
-    return enabled && available_providers.indexOf('ipns') > 0;
-  }
-
-  setNamesystemProvider() {
-    let code = `\nEmbarkJS.Names.setProvider('ipns', ${JSON.stringify(this.storageConfig.dappConnection[0] || {})});`;
-
-    let shouldInit = (namesystemConfig) => {
-      return (namesystemConfig.provider === 'ipns' && namesystemConfig.enabled === true);
-    };
-
-    this.embark.addProviderInit('names', code, shouldInit);
-  }
-
-
 }
 
 module.exports = IPFS;
