+++ lib/modules/swarm/embarkjs.js
@@ -103,3 +103,12 @@ __embarkSwarm.getUrl = function (hash) {
   return `${this._config.getUrl || this._connectUrl + '/bzz:/'}${hash}`;
 };
 
+const NotAvailable = "Not available with Swarm"
+
+__embarkSwarm.resolve = function () {
+  return new Error(NotAvailable);
+};
+
+__embarkSwarm.register = function () {
+  return new Error(NotAvailable);
+};
