+++ lib/modules/whisper/index.js
@@ -75,15 +75,13 @@ class Whisper {
     let connection = this.communicationConfig.connection || {};
 
     // todo: make the add code a function as well
-    let config = JSON.stringify({
+    let config = {
       server: canonicalHost(connection.host || defaultHost),
       port: connection.port || '8546',
       type: connection.type || 'ws'
-    });
+    };
 
-    config = JSON.stringify(config);
-
-    let code = "\nEmbarkJS.Messages.setProvider('whisper'," + config + ");";
+    let code = `\nEmbarkJS.Messages.setProvider('whisper', ${JSON.stringify(config)});`;
 
     let shouldInit = (communicationConfig) => {
       return (communicationConfig.provider === 'whisper' && communicationConfig.enabled === true);
